Moved project xml saving to data processor
